pyproject.toml
[tool.poetry]
name = "matricola-finder"
version = "0.1.0"
description = ""
authors = ["rikyiso01 <riky.isola@gmail.com>"]
readme = "README.md"
packages = [{include = "matricola_finder"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.24.1"
pyyaml = "^6.0"
bs4 = "^0.0.1"
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
